teller/requirements.txt
pdfplumber==0.5.23
python-dateutil==2.8.2
click
